Day 09: Rust version

+++ 2024/day-09/src/parser.rs
@@ -3,14 +3,14 @@ use winnow::{combinator::trace, stream::AsChar, PResult, Parser};
 #[derive(Clone, Debug)]
 pub struct PuzzleInput {
     // Remember to make the fields pub
-    pub disk: Vec<u8>,
+    pub disk: Vec<u64>,
 }
 
 pub fn part1(input: &str) -> PuzzleInput {
     PuzzleInput {
         disk: input
             .chars()
-            .filter_map(|c| c.to_digit(10).map(|d| d as u8) )
+            .filter_map(|c| c.to_digit(10).map(|d| d as u64))
             .collect(),
     }
 }
